worker/pyproject.toml
[project]
name = "image-worker"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
  "pika>=1.3.2",
  "pillow>=11.0.0",
  "psycopg[binary]>=3.2",
]

[tool.uv]
# optional: if you later add dev deps, uv can manage them too
